A minial version of Simplfied version of Core is finished. There are pretty many comments in the code, so I won't explain what it does in this commit message.
Tested it on two simple examples, where DLLs don't really
connect to some devices, instead, themselves emulate devices.

For testing purposes, I added to .gitignore the folder:
SimplifiedCore\DLLTests
So, create it if you want to play some.

And, yes, we've decided to get rid of that stupid idea of
using interfaces to build architecture of the project -
switching to classes soon.